Move components to ui and unify import paths
